contributors/ttwf_shenzhen/css-shapes/biqing/submitted/shape-outside-invalid-circle-000.html

<!DOCTYPE html>
<head>
    <meta charset="utf-8">
    <title>CSS Shapes Tests: Circle - invalid 1 arg</title>
    <link rel="author" title="biqing" href="biqing.kwok@qq.com">
    <link rel="help" href="http://www.w3.org/TR/css-shapes-1/#supported-basic-shapes">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <link rel="stylesheet" href="/resources/testharness.css">
    <meta name="flags" content="dom">
    <meta name="assert" content="shape-outside declaration is invalid if circle shape function only has one parameter">
    <style type="text/css">
    #float {
        shape-outside: circle(50px);
        float: left;
        width: 100px;
        height: 100px;
    }
    </style>
</head>
<body>
    <div id="log"></div>

    <div id="float"></div>

    <script>

    test(function() {
        assert_true(getComputedStyle(document.getElementById('float')).shapeOutside == 'auto');
    }, 'Circle - invalid 1 arg');

    </script>
</body>
</html>
